Add MiniMax H3 SPEED Sampler implementation and MIT License
- Introduced a self-contained MiniMax H3 SPEED Sampler node that integrates DCT spectral expansion, flow alignment, and adaptive stage execution.
- Implemented various DCT spectral primitives for signal processing.
- Added functions for flow and audio-video alignment math.
- Created a runtime execution method for the MiniMax H3 sampler.
- Defined an adaptive sampler node interface with preset mappings for different scaling strategies.
- Included an MIT License file for open-source distribution.
