+++ .gitignore
@@ -1,7 +1,9 @@
 __pycache__/
 *.py[cod]
+*$py.class
 .venv/
-.pytest_cache/
-*.egg-info/
-dist/
-build/
+venv/
+.vscode/
+.idea/
+*.log
+.DS_Store
