--- minimax_h3_speed/__init__.py
@@ -1,8 +1,0 @@
-"""minimax_h3_speed — core SPEED library for MiniMax-H3.
-
-Package layout:
-- config.py     — SpeedConfig, SCALE_PRESETS
-- flow.py       — transition math (scale_ratio alignment)
-- spectral.py   — DCT primitives
-- h3_runtime.py — run_progressive_stages (core loop)
-"""
