--- minimax_h3_speed/tests/test_sampler.py
@@ -1,486 +1,0 @@
-"""Minimal contract test for the single SPEED sampler node."""
-import importlib
-import math
-import sys
-from pathlib import Path
-from types import ModuleType
-
-import pytest
-import torch
-from minimax_h3_speed.config import SpeedConfig
-
-ROOT = Path(__file__).resolve().parents[1]
-sys.path.insert(0, str(ROOT))
-
-
-def _install_comfy_stubs():
-    comfy = ModuleType("comfy")
-    samplers = ModuleType("comfy.samplers")
-    utils = ModuleType("comfy.utils")
-    model_mgmt = ModuleType("comfy.model_management")
-    kdiff = ModuleType("comfy.k_diffusion")
-    ksampling = ModuleType("comfy.k_diffusion.sampling")
-    nested_tensor = ModuleType("comfy.nested_tensor")
-
-    class NestedTensor:
-        is_nested = True
-        def __init__(self, tensors):
-            self._tensors = tensors
-        def unbind(self):
-            return self._tensors
-    nested_tensor.NestedTensor = NestedTensor
-
-    samplers.sampler_object = lambda name: ("sampler", name)
-    utils.PROGRESS_BAR_ENABLED = True
-
-    def pack_latents(latents):
-        shapes, tensors = [], []
-        for t in latents:
-            shapes.append(list(t.shape))
-            tensors.append(t.reshape(t.shape[0], 1, -1))
-        return torch.cat(tensors, dim=-1), shapes
-
-    def unpack_latents(combined, shapes):
-        out, work = [], combined
-        for shape in shapes:
-            cut = math.prod(shape[1:])
-            out.append(work[:, :, :cut].reshape([work.shape[0]] + shape[1:]))
-            work = work[:, :, cut:]
-        return out
-
-    utils.pack_latents = pack_latents
-    utils.unpack_latents = unpack_latents
-    model_mgmt.intermediate_device = lambda: "cpu"
-
-    def sample_euler(model, x, sigmas, extra_args=None, callback=None, disable=None, **kwargs):
-        extra_args = {} if extra_args is None else extra_args
-        for i in range(len(sigmas) - 1):
-            sigma = sigmas[i]
-            denoised = model(x, sigma, **extra_args)
-            d = (x - denoised) / sigma
-            x = x + d * (sigmas[i + 1] - sigmas[i])
-            if callback is not None:
-                callback({"x": x, "i": i, "sigma": sigma, "denoised": denoised})
-        return x
-
-    ksampling.sample_euler = sample_euler
-    comfy.samplers = samplers
-    comfy.utils = utils
-    comfy.model_management = model_mgmt
-    comfy.k_diffusion = kdiff
-    comfy.k_diffusion.sampling = ksampling
-    comfy.nested_tensor = nested_tensor
-    sys.modules["comfy"] = comfy
-    for name, mod in [("samplers", samplers), ("utils", utils),
-                      ("model_management", model_mgmt),
-                      ("k_diffusion", kdiff), ("k_diffusion.sampling", ksampling),
-                      ("nested_tensor", nested_tensor)]:
-        sys.modules["comfy." + name] = mod
-
-
-# Install comfy stubs at module load so any test that imports `main` or
-# `minimax_h3_speed.h3_runtime` (which now import comfy at top level) works
-# without needing to call _install_comfy_stubs() first.
-_install_comfy_stubs()
-
-
-def test_node_exports_sampler():
-    _install_comfy_stubs()
-    mod = importlib.import_module("sampler_node")
-    cls = mod.MiniMaxH3SPEEDSampler
-    assert cls.RETURN_TYPES == ("LATENT", "LATENT")
-    assert cls.FUNCTION == "sample"
-    assert "MiniMaxH3SPEEDSampler" in mod.NODE_CLASS_MAPPINGS
-
-
-def test_input_schema_widgets_and_required_inputs():
-    _install_comfy_stubs()
-    mod = importlib.import_module("sampler_node")
-    inputs = mod.MiniMaxH3SPEEDSampler.INPUT_TYPES()
-    required = inputs["required"]
-    for key in ("noise", "guider", "sigmas", "latent_image",
-                "preset", "transition_mode"):
-        assert key in required, f"missing required input: {key}"
-    # delta_custom path is enabled with sigma-harvest calibration
-    assert "delta" in required
-    assert "power_A" in required
-    assert "power_beta" in required
-    assert "seed_offset" in required
-    assert required["preset"][0][0] == "half_then_full"
-
-
-def test_sample_runs_multi_stage():
-    _install_comfy_stubs()
-    from minimax_h3_speed.h3_runtime import run_progressive_stages
-
-    sample_calls = []
-
-    class FakeGuider:
-        model_patcher = type("MP", (), {"model": type("M", (), {
-            "sigma_shift_video": 12.0,
-            "sigma_shift_audio": 3.0,
-            "process_latent_out": lambda s, x: x,
-        })()})()
-        conds = {}
-
-        def sample(self, noise, latent_image, sampler, sigmas, callback=None, **kwargs):
-            sample_calls.append(len(sigmas))
-            if callback is not None:
-                callback(0, latent_image, latent_image, len(sigmas))
-            return latent_image
-
-    class FakeNoise:
-        seed = 42
-        def generate_noise(self, latent):
-            samples = latent.get("samples")
-            if getattr(samples, "is_nested", False):
-                vids = [s for s in samples.unbind() if s.ndim == 5]
-                auds = [s for s in samples.unbind() if s.ndim != 5]
-                return type("NT", (), {"is_nested": True, "unbind": lambda v=vids, a=auds: v + a})()
-            return samples
-
-    video = torch.zeros(1, 1, 2, 8, 8)
-    audio = torch.zeros(1, 1, 2, 44)
-
-    class FakeNested:
-        is_nested = True
-        def unbind(self):
-            return [video, audio]
-
-    nested = FakeNested()
-    latent = {"samples": nested}
-    sigmas = torch.linspace(1.0, 0.025, 20)
-    config = SpeedConfig(scales=(0.5, 1.0), transition_steps=(5,))
-
-    out, denoised = run_progressive_stages(
-        FakeNoise(), FakeGuider(), sigmas, latent, config,
-        sampler=type("S", (), {"name": "euler"})(),
-        nested_type=type("NT", (), {"is_nested": True})(),
-        disable_pbar=True, output_device=None,
-    )
-    assert out is not None
-    assert len(sample_calls) >= 2, f"expected ≥2 stages, got {len(sample_calls)}"
-
-
-def test_coupled_full_grid_noise_policy():
-    """coupled_full_grid: full-grid noise is shared across stages."""
-    _install_comfy_stubs()
-    from minimax_h3_speed.h3_runtime import run_progressive_stages
-
-    sample_calls = []
-    captured_noises = []
-
-    class FakeGuider:
-        model_patcher = type("MP", (), {"model": type("M", (), {
-            "sigma_shift_video": 12.0,
-            "sigma_shift_audio": 3.0,
-            "process_latent_out": lambda s, x: x,
-        })()})()
-        conds = {}
-
-        def sample(self, noise, latent_image, sampler, sigmas, callback=None, **kwargs):
-            sample_calls.append(len(sigmas))
-            if callback is not None:
-                callback(0, latent_image, latent_image, len(sigmas))
-            return latent_image
-
-    class FakeNoise:
-        seed = 77
-        def generate_noise(self, latent):
-            samples = latent.get("samples")
-            if getattr(samples, "is_nested", False):
-                vids = [s for s in samples.unbind() if s.ndim == 5]
-                auds = [s for s in samples.unbind() if s.ndim != 5]
-                nt = type("NT", (), {"is_nested": True, "unbind": lambda self: vids + auds})()
-                captured_noises.append(nt)
-                return nt
-            return samples
-
-    # Full-resolution latent: 32x32 spatial
-    video = torch.zeros(1, 1, 2, 32, 32)
-    audio = torch.zeros(1, 1, 2, 44)
-
-    class FakeNested:
-        is_nested = True
-        def unbind(self):
-            return [video, audio]
-
-    nested = FakeNested()
-    latent = {"samples": nested}
-    sigmas = torch.linspace(1.0, 0.025, 20)
-    config = SpeedConfig(scales=(0.5, 1.0), transition_steps=(5,), noise_policy="coupled_full_grid")
-
-    out, denoised = run_progressive_stages(
-        FakeNoise(), FakeGuider(), sigmas, latent, config,
-        sampler=type("S", (), {"name": "euler"})(),
-        nested_type=type("NT", (), {"is_nested": True})(),
-        disable_pbar=True, output_device=None,
-    )
-    assert out is not None
-    assert len(sample_calls) >= 2
-    assert len(captured_noises) > 0
-
-
-def test_aligned_speed_sigma_math():
-    """kappa = r / (1 + (r-1)q); t_tilde = kappa * q. Verify the paper formula."""
-    flow = importlib.import_module("minimax_h3_speed.flow")
-    for q, r in [(0.5, 2.0), (0.3, 2.0), (0.8, 4.0 / 3.0)]:
-        kappa, t = flow.aligned_speed_sigma(q, r)
-        assert abs(kappa - r / (1.0 + (r - 1.0) * q)) < 1e-6
-        assert abs(t - kappa * q) < 1e-12
-
-
-def test_resolve_transition_steps_explicit_example():
-    """Explicit mode places transitions at the flat per-scale default (5)."""
-    mod = importlib.import_module("sampler_node")
-    h3_runtime = importlib.import_module("minimax_h3_speed.h3_runtime")
-    sigmas = torch.linspace(1.0, 0.025, 20)
-    explicit_steps = h3_runtime.resolve_transition_steps(
-        SpeedConfig(scales=(0.5, 1.0), transition_steps=(5,), transition_mode="explicit"),
-        sigmas, 64, 64,
-    )
-    assert tuple(explicit_steps) == (5,)
-
-
-def test_active_av_shifts_returns_audio_scale_from_ratio():
-    """_active_av_shifts should compute audio_scale = video_shift / audio_shift,
-    not read a non-existent 'audio_scale' attribute.
-
-    With shift_video=12.0 and shift_audio=3.0, the correct audio_scale is 4.0.
-    """
-    from minimax_h3_speed.h3_runtime import _active_av_shifts
-
-    class FakeGuider:
-        model_patcher = type("MP", (), {"model": type("M", (), {
-            "sigma_shift_video": 12.0,
-            "sigma_shift_audio": 3.0,
-            "diffusion_model": type("DM", (), {
-                "sigma_shift_video": 12.0,
-                "sigma_shift_audio": 3.0,
-            })(),
-        })()})()
-
-    v_shift, a_shift, a_scale = _active_av_shifts(FakeGuider())
-    assert v_shift == 12.0
-    assert a_shift == 3.0
-    assert abs(a_scale - 4.0) < 1e-6
-
-
-def test_audio_scale_is_ratio_not_one():
-    """Verify audio_scale = video_shift / audio_shift, not 1.0."""
-    from minimax_h3_speed.h3_runtime import _active_av_shifts
-
-    class FakeGuider:
-        model_patcher = type("MP", (), {"model": type("M", (), {
-            "sigma_shift_video": 12.0,
-            "sigma_shift_audio": 3.0,
-        })()})()
-
-    _, _, a_scale = _active_av_shifts(FakeGuider())
-    assert abs(a_scale - 4.0) < 1e-6
-
-
-def test_sigma_policy_canonical_vs_no_alignment():
-    """canonical: apply kappa alignment; no_alignment: no rescaling."""
-    from minimax_h3_speed.h3_runtime import run_progressive_stages
-
-    canonical_calls = []
-    no_align_calls = []
-
-    def make_guider(calls_list):
-        class FakeGuider:
-            model_patcher = type("MP", (), {"model": type("M", (), {
-                "sigma_shift_video": 12.0,
-                "sigma_shift_audio": 3.0,
-                "process_latent_out": lambda s, x: x,
-            })()})()
-            conds = {}
-            def sample(self, noise, latent_image, sampler, sigmas, callback=None, **kwargs):
-                calls_list.append(len(sigmas))
-                if callback is not None:
-                    callback(0, latent_image, latent_image, len(sigmas))
-                return latent_image
-        return FakeGuider()
-
-    video = torch.zeros(1, 1, 2, 8, 8)
-    audio = torch.zeros(1, 1, 2, 44)
-    nested = type("NT", (), {"is_nested": True, "unbind": lambda self: [video, audio]})()
-    latent = {"samples": nested}
-    sigmas = torch.linspace(1.0, 0.025, 20)
-
-    config_canon = SpeedConfig(scales=(0.5, 1.0), transition_steps=(5,), sigma_policy="canonical")
-    config_noalign = SpeedConfig(scales=(0.5, 1.0), transition_steps=(5,), sigma_policy="no_alignment")
-
-    run_progressive_stages(
-        type("N", (), {"seed": 42, "generate_noise": lambda s, l: l["samples"]})(),
-        make_guider(canonical_calls),
-        sigmas, latent, config_canon,
-        sampler=type("S", (), {"name": "euler"})(),
-        nested_type=type("NT", (), {"is_nested": True})(),
-        disable_pbar=True,
-    )
-    run_progressive_stages(
-        type("N", (), {"seed": 42, "generate_noise": lambda s, l: l["samples"]})(),
-        make_guider(no_align_calls),
-        sigmas, latent, config_noalign,
-        sampler=type("S", (), {"name": "euler"})(),
-        nested_type=type("NT", (), {"is_nested": True})(),
-        disable_pbar=True,
-    )
-    # Both should run the same number of stages
-    assert len(canonical_calls) == len(no_align_calls)
-    assert len(canonical_calls) >= 2
-
-
-def test_invalid_transition_mode_raises():
-    """Config with unsupported transition_mode must fail fast."""
-    with pytest.raises(ValueError, match="transition_mode"):
-        SpeedConfig(scales=(0.5, 1.0), transition_steps=(5,), transition_mode="unknown")
-
-
-def test_invalid_noise_policy_raises():
-    with pytest.raises(ValueError, match="noise_policy"):
-        SpeedConfig(scales=(0.5, 1.0), transition_steps=(5,), noise_policy="evil")
-
-
-def test_single_scale_must_be_full():
-    with pytest.raises(ValueError, match="single scale must be 1.0"):
-        SpeedConfig(scales=(0.5,), transition_steps=())
-
-
-def test_final_scale_must_be_full():
-    with pytest.raises(ValueError, match="final scale must be 1.0"):
-        SpeedConfig(scales=(0.5, 0.75), transition_steps=(3,))
-
-
-def test_transition_steps_count_matches_scales():
-    with pytest.raises(ValueError, match="need .* transition steps"):
-        SpeedConfig(scales=(0.5, 0.75, 1.0), transition_steps=(5,))
-
-
-def test_reentry_noise_formula():
-    """reentry_noise(internal, start_sigma) = internal / start_sigma."""
-    from minimax_h3_speed.flow import reentry_noise
-    internal = torch.tensor([1.0, 2.0, 3.0])
-    result = reentry_noise(internal, 0.5)
-    assert torch.allclose(result, internal / 0.5)
-
-
-def test_reentry_noise_raises_on_zero():
-    from minimax_h3_speed.flow import reentry_noise
-    with pytest.raises(ValueError, match="start_sigma must be positive"):
-        reentry_noise(torch.zeros(3), 0.0)
-
-
-def test_kappa_formula():
-    """κ = r / (1 + (r-1)t) per Eq. (5)."""
-    from minimax_h3_speed.flow import aligned_speed_sigma
-    r = 2.0
-    t = 0.5
-    kappa, new_q = aligned_speed_sigma(t, r)
-    expected_kappa = r / (1.0 + (r - 1.0) * t)
-    assert abs(kappa - expected_kappa) < 1e-10
-
-
-def test_time_shift_sigma_raises_on_bad_inputs():
-    """time_shift_sigma must reject non-positive shifts."""
-    from minimax_h3_speed.flow import time_shift_sigma
-    with pytest.raises(ValueError):
-        time_shift_sigma(0.5, 0.0, 1.0)
-    with pytest.raises(ValueError):
-        time_shift_sigma(0.5, 1.0, -1.0)
-
-
-def test_time_shift_sigma_identity_at_full_res():
-    """At q = q_ref (no shift), returns q unchanged."""
-    from minimax_h3_speed.flow import time_shift_sigma
-    result = time_shift_sigma(0.5, 1.0, 1.0)
-    assert abs(result - 0.5) < 1e-10
-
-
-def test_flow_time_shift_sigma():
-    """time_shift_sigma bridges video→audio sigma space correctly."""
-    from minimax_h3_speed.flow import time_shift_sigma
-    # With video_shift=2.0, audio_shift=1.0, q_video=0.5:
-    #   base = 0.5 / (2 + 0.5 * (1-2)) = 0.5 / 1.5 = 1/3
-    #   q_audio = 1.0 * (1/3) / (1 + 0 * (1/3)) = 1/3
-    q_audio = time_shift_sigma(0.5, 2.0, 1.0)
-    assert abs(q_audio - (1.0/3.0)) < 1e-6
-
-
-def test_resolve_transition_steps_delta_custom_matches_recommend():
-    """Given config with transition_mode='delta_custom', the resolved steps
-    must equal recommend_configs output for the same parameters."""
-    from minimax_h3_speed.h3_runtime import resolve_transition_steps
-    from minimax_h3_speed.harvest import recommend_configs
-    sigmas = torch.linspace(1.0, 0.0, 21)  # 20 steps
-    config = SpeedConfig(
-        scales=(0.5, 1.0),
-        transition_steps=(5,),
-        transition_mode="delta_custom",
-        delta=0.01,
-        power_A=219.48,
-        power_beta=2.42,
-        full_latent_h=45,
-        full_latent_w=80,
-    )
-    resolved = resolve_transition_steps(config, sigmas, H_full=45, W_full=80)
-    rec = recommend_configs(219.48, 2.42, sigmas, latent_h=45, latent_w=80)
-    expected = rec["half_then_full"]["transition_steps"]
-    assert resolved == tuple(expected)
-
-
-def test_resolve_transition_steps_explicit_ignores_delta():
-    """For 'explicit' mode, resolved steps must equal config.transition_steps,
-    regardless of delta/power_A/power_beta values."""
-    from minimax_h3_speed.h3_runtime import resolve_transition_steps
-    sigmas = torch.linspace(1.0, 0.0, 21)
-    config = SpeedConfig(
-        scales=(0.5, 1.0),
-        transition_steps=(7,),
-        transition_mode="explicit",
-        delta=0.5,  # irrelevant in explicit mode
-        power_A=999.0,
-        power_beta=9.0,
-        full_latent_h=45,
-        full_latent_w=80,
-    )
-    resolved = resolve_transition_steps(config, sigmas)
-    assert resolved == (7,)
-
-
-def test_resolve_transition_steps_validation():
-    """Transition steps must be in (0, len(sigmas)-1)."""
-    from minimax_h3_speed.h3_runtime import resolve_transition_steps
-    sigmas = torch.tensor([1.0, 0.5, 0.0])
-    config = SpeedConfig(
-        scales=(0.5, 1.0),
-        transition_steps=(1,),  # valid for config (>= 1)
-        transition_mode="explicit",
-        delta=0.01,
-        power_A=219.48,
-        power_beta=2.42,
-        full_latent_h=45,
-        full_latent_w=80,
-    )
-    resolved = resolve_transition_steps(config, sigmas)
-    assert resolved == (1,)
-
-
-def test_activation_time_matches_canonical_formula():
-    """Verify activation_time against hand-computed values from Eq. 9."""
-    from minimax_h3_speed.h3_runtime import activation_time
-    import math
-    result = activation_time(100.0, 0.01)
-    expected = 1.0 / (1.0 + math.sqrt(0.01 / (100.0 * (101.0 - 0.01))))
-    assert abs(result - expected) < 1e-10
-
-
-def test_flow_time_shift_sigma():
-    """time_shift_sigma bridges video→audio sigma space correctly."""
-    from minimax_h3_speed.flow import time_shift_sigma
-    # With video_shift=2.0, audio_shift=1.0, q_video=0.5:
-    #   base = 0.5 / (2 + 0.5 * (1-2)) = 0.5 / 1.5 = 1/3
-    #   q_audio = 1.0 * (1/3) / (1 + 0 * (1/3)) = 1/3
-    q_audio = time_shift_sigma(0.5, 2.0, 1.0)
-    assert abs(q_audio - (1.0/3.0)) < 1e-6
