+++ pyproject.toml
@@ -3,17 +3,13 @@ requires = ["setuptools>=61", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "comfyui-mini-max-h3-speed-sampler"
-version = "0.1.0"
-description = "WIP — SPEED progressive-resolution sampler as a ComfyUI SAMPLER node (single-node MVP)."
+name = "comfyui-minimax-h3-speed"
+version = "1.0.0"
+description = "Adaptive SPEED progressive-resolution sampler for MiniMax-H3 in ComfyUI (Supports Turbo LoRAs & Standard Schedules)."
 readme = "README.md"
-license = {text = "PolyForm Noncommercial 1.0.0"}
+license = {text = "MIT"}
 requires-python = ">=3.10"
-# Requires ComfyUI and torch at runtime; these are provided by the host.
-dependencies = ["torch"]
-
-[project.entry-points."comfyui.custom_nodes"]
-comfyui_mini_max_h3_speed_sampler = "sampler"
+dependencies = ["torch>=2.0"]
 
 [tool.setuptools.packages.find]
 include = ["*"]
